cleanup processor code
