+++ src/Processors/VendorPropertyValidation.php
@@ -8,7 +8,7 @@ use Radebatz\OpenApi\Routing\Annotations\MiddlewareProperty;
 use Radebatz\OpenApi\Routing\RoutingAdapterInterface;
 
 /**
- * OpenApi processor to allow easy validation of vendor properties.
+ * Validation of vendor (this)  properties.
  */
 class VendorPropertyValidation
 {
